requirements/common.txt
alembic==1.8.0
async-generator==1.10
attrs==21.4.0
bleach==5.0.1
blinker==1.4
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
cryptography==37.0.4
defusedxml==0.7.1
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.1
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-PageDown==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-SSLify==0.1.5
Flask-WTF==1.0.1
h11==0.13.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
Markdown==3.3.7
MarkupSafe==2.1.1
multidict==6.0.2
outcome==1.2.0
pycparser==2.21
pyOpenSSL==22.0.0
PyNaCl==1.5.0
PySocks==1.7.1
PyYAML==6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests-toolbelt==0.9.1
rich==12.4.4
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
trio==0.21.0
trio-websocket==0.9.2
visitor==0.1.3
webencodings==0.5.1
Werkzeug==2.1.2
wheel==0.37.1
wsproto==1.1.0
WTForms==3.0.1
zipp==3.8.0
